backend/complex/save-notes/requirements.txt
grpcio==1.62.0
protobuf==4.25.3
python-dotenv==1.0.1
